Modifications for 64-bit OS (#218)
* Modifications for 64-bit OS

* Change unsigned long to uint32_t in examples

* Modify printf in example: Change %lu to %u

* run clang-format v12.0.1 on unchanged some files

---------

+++ examples_RPi/helloworld_rx.cpp
@@ -29,8 +29,8 @@ const uint16_t other_node = 01;
 
 struct payload_t
 { // Structure of our payload
-    unsigned long ms;
-    unsigned long counter;
+    uint32_t ms;
+    uint32_t counter;
 };
 
 int main(int argc, char** argv)
@@ -56,7 +56,7 @@ int main(int argc, char** argv)
             payload_t payload;
             network.read(header, &payload, sizeof(payload));
 
-            printf("Received payload: counter=%lu, origin timestamp=%lu\n", payload.counter, payload.ms);
+            printf("Received payload: counter=%u, origin timestamp=%u\n", payload.counter, payload.ms);
         }
         //sleep(2);
         delay(2000);
